libs/utils(refactor): reorganize tabulation utils
no significant changes here other than a test tweak due to switching fixtures
