+++ libs/utils/src/tabulation/contest_filtering.ts
@@ -4,42 +4,30 @@ import {
   Election,
   ElectionDefinition,
   PrecinctId,
-  Tabulation,
   AnyContest,
   Contests,
 } from '@votingworks/types';
 import { assert } from '@votingworks/basics';
 import {
   createElectionMetadataLookupFunction,
-  getBallotStylesByPartyId,
-  getBallotStylesByPrecinctId,
   getContestById,
 } from './lookups';
-import { doesContestAppearOnPartyBallot } from './election_utils';
 
-export function unionSets<T>(sets: Array<Set<T>>): Set<T> {
-  const combinedSet = new Set<T>();
-  for (const set of sets) {
-    for (const item of set) {
-      combinedSet.add(item);
-    }
-  }
-  return combinedSet;
-}
-
-export function intersectSets<T>(sets: Array<Set<T>>): Set<T> {
-  const intersectionSet = new Set<T>();
-  if (sets.length === 0) {
-    return intersectionSet;
-  }
-  const firstSet = sets[0];
-  assert(firstSet);
-  for (const item of firstSet) {
-    if (sets.every((set) => set.has(item))) {
-      intersectionSet.add(item);
-    }
-  }
-  return intersectionSet;
+/**
+ * Contests appear on ballots or not based on the district the contest is
+ * associated with and the party. This function just covers the party part. Rules:
+ *   - ballot measures can appear on ballots of any party
+ *   - candidates contests with an associated party can only appear on ballots of the same party
+ */
+export function doesContestAppearOnPartyBallot(
+  contest: AnyContest,
+  ballotPartyId?: string
+): boolean {
+  return (
+    contest.type === 'yesno' ||
+    !contest.partyId ||
+    contest.partyId === ballotPartyId
+  );
 }
 
 function buildBallotStyleContestIdsLookup(
@@ -62,24 +50,6 @@ function buildBallotStyleContestIdsLookup(
 
 export const getContestIdsForBallotStyle = createElectionMetadataLookupFunction(
   buildBallotStyleContestIdsLookup
-);
-
-function buildPartyContestIdsLookup(
-  election: Election
-): Record<string, Set<ContestId>> {
-  const lookup: Record<string, Set<ContestId>> = {};
-  for (const party of election.parties) {
-    lookup[party.id] = new Set(
-      election.contests
-        .filter((c) => doesContestAppearOnPartyBallot(c, party.id))
-        .map((c) => c.id)
-    );
-  }
-  return lookup;
-}
-
-export const getContestIdsForParty = createElectionMetadataLookupFunction(
-  buildPartyContestIdsLookup
 );
 
 function buildPrecinctContestIdsLookup(
@@ -112,90 +82,6 @@ export const getContestIdsForPrecinct = createElectionMetadataLookupFunction(
   buildPrecinctContestIdsLookup
 );
 
-export function getBallotStyleIdsForFilter(
-  electionDefinition: ElectionDefinition,
-  filter?: Tabulation.Filter
-): Set<BallotStyleId> {
-  const { election } = electionDefinition;
-
-  let ballotStyleIds = new Set(election.ballotStyles.map((bs) => bs.id));
-  if (!filter) return ballotStyleIds;
-
-  // Ballot Style, Party, and Precinct filters can all narrow down contests
-
-  // narrow down by explicit Ballot Style filter
-  if (filter.ballotStyleIds) {
-    ballotStyleIds = intersectSets([
-      ballotStyleIds,
-      new Set(filter.ballotStyleIds),
-    ]);
-  }
-
-  // narrow down by Party filter
-  if (filter.partyIds) {
-    const { partyIds } = filter;
-    const ballotStyleIdsRestrictedByParty: Set<BallotStyleId> = unionSets(
-      partyIds.map((partyId) => {
-        const ballotStyles = getBallotStylesByPartyId(
-          electionDefinition,
-          partyId
-        );
-        return new Set(ballotStyles.map((bs) => bs.id));
-      })
-    );
-
-    ballotStyleIds = intersectSets([
-      ballotStyleIds,
-      ballotStyleIdsRestrictedByParty,
-    ]);
-  }
-
-  // narrow down by Precinct filter
-  if (filter.precinctIds) {
-    const { precinctIds } = filter;
-    const ballotStyleIdsRestrictedByPrecinct: Set<BallotStyleId> = unionSets(
-      precinctIds.map((precinctId) => {
-        const ballotStyles = getBallotStylesByPrecinctId(
-          electionDefinition,
-          precinctId
-        );
-        return new Set(ballotStyles.map((bs) => bs.id));
-      })
-    );
-
-    ballotStyleIds = intersectSets([
-      ballotStyleIds,
-      ballotStyleIdsRestrictedByPrecinct,
-    ]);
-  }
-
-  return ballotStyleIds;
-}
-
-function getContestIdsForBallotStyleIds(
-  electionDefinition: ElectionDefinition,
-  ballotStyleIds: BallotStyleId[]
-): Set<ContestId> {
-  return unionSets(
-    ballotStyleIds.map((ballotStyleId) =>
-      getContestIdsForBallotStyle(electionDefinition, ballotStyleId)
-    )
-  );
-}
-
-/**
- * Filters can restrict the ballot styles that can appear in a report - if there
- * are no ballot styles in a report, it's not a valid report.
- */
-export function getContestIdsForFilter(
-  electionDefinition: ElectionDefinition,
-  filter?: Tabulation.Filter
-): Set<ContestId> {
-  return getContestIdsForBallotStyleIds(electionDefinition, [
-    ...getBallotStyleIdsForFilter(electionDefinition, filter),
-  ]);
-}
-
 export function mapContestIdsToContests(
   electionDefinition: ElectionDefinition,
   contestIds: Set<ContestId>
@@ -203,51 +89,6 @@ export function mapContestIdsToContests(
   return [...contestIds].map((contestId) =>
     getContestById(electionDefinition, contestId)
   );
-}
-
-export function convertGroupSpecifierToFilter(
-  group: Tabulation.GroupSpecifier
-): Tabulation.Filter {
-  return {
-    ballotStyleIds: group.ballotStyleId ? [group.ballotStyleId] : undefined,
-    partyIds: group.partyId ? [group.partyId] : undefined,
-    precinctIds: group.precinctId ? [group.precinctId] : undefined,
-    scannerIds: group.scannerId ? [group.scannerId] : undefined,
-    batchIds: group.batchId ? [group.batchId] : undefined,
-    votingMethods: group.votingMethod ? [group.votingMethod] : undefined,
-  };
-}
-
-export function mergeFilters(
-  filter1: Tabulation.Filter,
-  filter2: Tabulation.Filter
-): Tabulation.Filter {
-  return {
-    ballotStyleIds:
-      filter1.ballotStyleIds || filter2.ballotStyleIds
-        ? [...(filter1.ballotStyleIds || []), ...(filter2.ballotStyleIds || [])]
-        : undefined,
-    partyIds:
-      filter1.partyIds || filter2.partyIds
-        ? [...(filter1.partyIds || []), ...(filter2.partyIds || [])]
-        : undefined,
-    precinctIds:
-      filter1.precinctIds || filter2.precinctIds
-        ? [...(filter1.precinctIds || []), ...(filter2.precinctIds || [])]
-        : undefined,
-    scannerIds:
-      filter1.scannerIds || filter2.scannerIds
-        ? [...(filter1.scannerIds || []), ...(filter2.scannerIds || [])]
-        : undefined,
-    batchIds:
-      filter1.batchIds || filter2.batchIds
-        ? [...(filter1.batchIds || []), ...(filter2.batchIds || [])]
-        : undefined,
-    votingMethods:
-      filter1.votingMethods || filter2.votingMethods
-        ? [...(filter1.votingMethods || []), ...(filter2.votingMethods || [])]
-        : undefined,
-  };
 }
 
 export function getContestsForPrecinct(
@@ -261,18 +102,4 @@ export function getContestsForPrecinct(
 
   const contestIds = getContestIdsForPrecinct(electionDefinition, precinctId);
   return mapContestIdsToContests(electionDefinition, contestIds);
-}
-
-/**
- * Currently, if results are split by batch and scanner then they are opportunistic,
- * only including non-zero splits.
- */
-export function groupBySupportsZeroSplits(
-  groupBy: Tabulation.GroupBy
-): boolean {
-  if (groupBy.groupByBatch || groupBy.groupByScanner) {
-    return false;
-  }
-
-  return true;
 }
