+++ libs/utils/src/tabulation/index.ts
@@ -1,4 +1,5 @@
+export * from './arguments';
 export * from './contest_filtering';
-export * from './election_utils';
 export * from './lookups';
 export * from './tabulation';
+export * from './transformations';
